+++ client/hifisampler.csproj
